// projects/web-app/src/types/auction-types.ts
export type AuctionCard = {
  id: string; // オークションのID
  current_highest_bid: number; // 最高額
  end_time: Date; // オークションの終了時刻
  start_time: Date; // オークションの開始時刻
  status: TaskStatus;
  task: string; // タスクのタイトル
  detail: string | null; // タスクの詳細
  image_url: string | null; // taskの画像
  category: string | null; // カテゴリ
  group_id: string; // グループのID
  group_name: string; // グループの名前
  bids_count: number; // 入札数
  is_watched: boolean; // オークションがウォッチリストに登録されているかどうか
  score: number | null; // pgroonga_score
  task_highlighted: string | null; // タスクのタイトルのハイライト
  detail_highlighted: string | null; // タスクの詳細のハイライト
  executors_json:
    | {
        id: string | null; // listのidとして使用
        rating: number | null; // Executorの評価
        userId: string | null; // listのidとして使用
        userImage: string | null; // taskExecutorのicon画像
        userSettingsUsername: string | null; // app内で表示される名前
      }[]
    | string;
};